ecfr-analytics/api/pyproject.toml
[project]
name = "ecfr-analytics-api"
version = "0.1.0"
description = "FastAPI service for eCFR analytics"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.112.2",
    "uvicorn==0.30.6",
    "google-cloud-bigquery==3.25.0",
    "python-dotenv==1.0.1",
]

[tool.uv]
dev-dependencies = []